tests/wpt/tests/css/css-box/margin-trim/computed-margin-values/grid-block-end-item-spans-multiple-rows.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-grid">
<meta name="assert" content="item that spans into last row should have block-end margin trimmed">
</head>
<style>
grid {
    display: grid;
    border: 1px solid black;
    grid-template-columns: auto auto;
    margin-trim: block-end;
}
item {
    display: block;
    width: 50px;
    height: 50px;
    margin-bottom: 10px;
}
.row-two {
    grid-row: 2;
    background-color: green;
}
.span-two-rows {
    grid-row: span 2;
    background-color: blue;
    height: 90px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('grid > item')">
    <div id="target">
    <grid>
        <item data-expected-margin-bottom="0" class="span-two-rows"></item>
        <item data-expected-margin-bottom="0" class="row-two"></item>
    </grid>
    </div>
</body>
</html>
