third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/grid-block-end-item-spans-multiple-rows-expected.txt

This is a testharness.js-based test.
FAIL grid > item 1 assert_equals: 
<item data-expected-margin-bottom="0" class="span-two-rows"></item>
margin-bottom expected "0" but got "10"
FAIL grid > item 2 assert_equals: 
<item data-expected-margin-bottom="0" class="row-two"></item>
margin-bottom expected "0" but got "10"
Harness: the test ran to completion.
